Do not resolve groups id to check reablility of shadow db
Calling an NSS function (in this case groupIds) inside an NSS module is
garanteed to fail or lead to unpredictable results. So, to check if
the user belongs to group shadow and is able to read shadow database, we
know do a simple file open test instead of checking group membership.

Co-authored-by: Didier Roche <didrocks@ubuntu.com>