Fix segfault while iterating on object with no database
If the cache database couldn’t be created for some reasons, we were then
trying to close the whole cache on a nil pointer.
Now, return early to avoid this segfault.

Note that this will change with the new open/close cache logic we will
introduce in the next days.