Parse additional PAM options when building with integrationtests tags
This tag is only used for integration tests. It allows to control the mocks
only with the build configuration via local values.
The net advantage is that this code, which is poking security holes as we can
have the PAM module not running as root and save the cache in temporary
directory is never shipped in production.