Add Debian packaging for aad-auth
Build libpam-aad and libnss-aad from the aad-auth source package,
representing the PAM and NSS modules. There will be a follow-up
aad-auth-cli package that will include the CLI tool.

Config files were also added for pam-auth-update and an example for
aad.conf. An nsswitch postinst configuration will be added in a
subsequent pull request.

Co-authored-by: denisonbarbosa <denison.barbosa@canonical.com>