Avoid opening a syslog connection in tests
Previously we passed a custom log writer when creating a syslog logger,
but even with that we first tried to open an actual syslog connection
before overriding it with our custom writer.

This appears to be an issue on the LP builders where syslog is not
active.

To fix this, when creating a new syslog logger, first apply the given
options and only open a syslog connection if a writer wasn't already
set.