Adding comments to ignore some linter errors on unsafe calls
Since those errors were on tests and only existed because Go understands
that a C type of a package is different than the same type include in
another package (hence why we had to use the unsafe.Pointer() in order
to trick the compiler to ignore this difference), shouldn't be a problem
ignoring those checks inside the test functions.
Please let me know if you guys disagree.