Fix tests failing due to timezone differences
LastOnlineAuth is serialized as a time.Time object, and its string
representation varies depending on the timezone.

To fix this we can change the ini stringifier to return the unix time
for an user.

For the user set tests we compared UserRecord objects which were
serialized/deserialized to yaml. We can just compare the changed
attributes in this case.