Refactoring the existing setup for the cache in pam integration tests
Modified the pam integration tests structure to no longer rely on .db
files for the testing. Now, the sql required for the creation of the
databases is provided, along with the dump files for test cases that
require a populated database.

During the tests, the databases are created in a temp dir and populated
with the data stored in the dump files.