Add script to generate completions and manpage
Cobra doesn't have a way to generate a single manpage for the command
and its subcommands, so the result gets cluttered with manpages for the
root command or the user/config commands which have no purpose other
than being a stepping stone for the child commands.

Mango helps us by generating a single manpage for the whole thing which
is much more user-friendly and easier to parse.

The completion and generators code is mostly taken from adsys, with
small changes to account for the newer way of generating manpages.