Updating some golden files
Since now we don't use the function to sort the users that are inserted
in the db in lexicographic order, I had to update the dumps to make sure
they would be inserted this way and update the uid and gid of some
users.
The golden files were updated to reflect the changes.