Parse config by mapping ini to struct
The go-ini package provides its own way of "unmarshaling" ini structures
by calling `MapTo` on sections. We can use this to build the config
struct.

This also makes the change of failing hard when we encounter type
mismatch errors, such as string values for the
`offline_credentials_expiration` field - which I don't believe we should
support.