Tests for EntryIterations are no longer executed in parallel.
(passwd|group|shadow)EntryIteration is a global state that controls the
iterator position of an ongoing iteration. When trying to run tests that use
StartEntryIteration in parallel, those tests will fail (as this global
state will be shared along threads) in cases they're not supposed to.
Since this global state is the expected behavior of the API, tests that
rely on it should not be executed in parallel.