Add option to move user home directory
When updating a home directory, the user can specify --move-home or -m
in order to also attempt moving the home directory contents from the old
path to the new path.

Implementation aims to follow usermod behavior as much as possible:
- passwd entry is _always_ updated regardless of the result of the move
  action
- move action fails if target directory exists
- move action fails if source directory does not exist
- move action fails if source directory is a file

We attempt the move with os.Rename at first in case the source and
target are on the same filesystem, and if that fails we fall back to the
`mv` command.

Some additional flag validation is also added to ensure correct usage of
the flag.