Do not move homedir if the user has open processes
If --move-home is requested, check if the user has open processes before
attempting to change and move the homedir.

This is accomplished by checking /proc for processes/threads owned by
the target user. If such a process exists, the program will stop at the
first occurence and return an error indicating the problematic PID,
without making any changes to the passwd database and filesystem.

Additionally, the program skips over processes that are running in a
different chroot.

The /proc path is configurable in order to aid with testing.