Fixing bugs in the cache tests
Some of the test cases for the cache tests were not using any database
during the tests, but they should use, since the objective of those test
cases is to query a filled database rather than not having any data at
all.

Also, a test case in test_build was misleading and was testing a different
occurence, since we don't create the cache ever in NSS.
Now, after the renaming, it shows better what is being tested.