Updating Rust packaging jobs in the action
We spotted that the job responsible for updating the fake checksum file
was also adding node modules to our repository if the action was
executed. To prevent that, we decided to rewrite it in shell, which
ended up simplifying the workflow while also avoid unnecessary
dependencies and changes.

Since the job was rewritten in shell and it is related to the vendored
sources job as well, we decided to also join both in a single one called
update-rust-packaging. This cuts our action time substantially and
reduces the ammount of commits pushed to main in case both steps make
changes.