Replacing backend used by rustix
The rustix crate has the option of two backends to use: libc and
linux_raw.
The linux_raw backend is shipped with .a files, which are the outlines
that are statically linked at build time. Those files get removed when
building the package and then "Houston, we have a problem".
Usually this doesn't matter since the default way rustix builds itself
is using asm, but this feature is experimental for ppc64el and MIPS
architectures.
To fix this problem, we need to switch to using the libc backend, which
is composed of only source files and, therefore, represents no issue to
us.