Switch the cargo config patch to the right branch
In https://github.com/ubuntu/aad-auth/pull/184, we changed the branch
used as the source for the libnss crate. Sadly, we forgot to also update
the config.toml patch that replaced that source when building the
package, which ended up breaking the build.

This commit fixes the target to be the right branch.