Shadow mode eval now considers effective ids
In order to determine the shadow mode of the cache, we were using
unix.Access() which considered only the real uid and gid of the current
process. This was causing sudo calls to fail, since they only change the
effective uid of the current process.

Replacing unix.Access() by unix.Faccessat() allows us to consider the
effective uid of the caller instead and allowing sudo calls to work, if
the user is in the sudo group or in the sudoers file.