Apply issue template changes to the repo
With https://github.com/canonical/desktop-engineering/pull/2, we defined
some yaml forms to be used as issue templates for reporting bugs and
requesting features on GitHub.

This provides a better and more defined way for users to interact with
us and provide us with the relevant information we need to solve the
problem.